update ChannelDailyCog to be more consistent with other cogs
For one, it was fetching an unnecessary score before calling
ScoresCog.update_scores(), so that was changed accordingly.

All guild daily tables are also created when the bot starts, similar
to the score & game night tables. This means that their existence no
longer has to be checked for via sqlite_master (e.g. in
clear_daily_claims).