Store seconds rather than minutes during game nights
This avoids the problem of 59 seconds being truncated to 0 minutes
(or any amount of seconds less than a minute). Total time spent
in the summary is still sent in H:MM format.

Also addressed case where if a user didn't meet any point
thresholds for a game night, the max function would throw a
ValueError.