Implement bonus role cog
It handles giving a special role to the top 12 users within a guild,
as well as providing a 5 point bonus every time a user gains that
role.

To avoid possible infinite recursion, the update_scores function in
the Scores cog now has the update_roles argument, which controls what
it sounds like.